feat: upgrade vac tower and extract superclass
